feat: persist spentbook log, add ultraman
Change 1: enable a configurable, fixed local port

Previously the local port was chosen randomly.  This made it
a pain to invoke multiple spentbook nodes because they always
varied between invocations.

With this change, the default port is 1111.  Also, there is a
--port arg to enable a custom port.

Change 2: persist spentbook log

The wallet already stores a wallet.dat file to persist between
executions however the spentbook nodes were not. This
asymmetry meant that wallet and spentbook would get out of
sync if the spentbook nodes were restarted.

With this commit, the spentbook now writes serialized (via serde/ron)
entries (one per line) to an append-only log file.  RON is used
instead of bincode because we want one entry per line, and bincode
serialization includes newlines, but RON does not.

The logfile is read upon startup and each entry is replayed
into memory via SpentbookNodeMock::log_spent()

The path of the logfile is provided via a new required
argument: spentbook-file.

Change 3: add ultraman process manager.

Ultraman is a rust port of the `foreman` process manager from
the ruby ecosystem.  It processes a Procfile and makes it easy
to start, monitor output, and stop multiple processes.

We use it to start 3 spentbook node processes, each on a different port.

So all the user has to do is: `ultraman start` and a 3-node spentbook
section starts running.  To end it, just ctrl-c the ultraman process.

Related to this, there is a new shell script for the purpose of deleting
(reset) both the wallet and spentbook files. All processes should be
stopped before doing this.

Also there is a new Procman file, which is read by ultraman.

Change 4: add README.md file.

The readme has build and usage info, as well as examples of using the
wallet.

spentbook changes:
* add cargo dep on ron 0.7.0
* create struct SpentLogEntry for serialization
* add --port arg, with default 1111
* add --spentbook-file arg (required)
* write to log file after ::log_spent() succeeds
* read in from log file once SpentBookNodeMock is created.

wallet changes:
* set default for --join-spentbook to 127.0.0.1:1111
* add wallet-file arg with default .wallet.dat